refactor: change to include credentials
