+++ next.config.js
@@ -4,22 +4,6 @@ const nextConfig = {
   experimental: {
     outputStandalone: true,
   },
-  async rewrites() {
-    return [
-      {
-        source: "/api/auth/sign-in",
-        destination: "https://skylab-backend.herokuapp.com/api/auth/sign-in",
-      },
-      {
-        source: "/api/auth/info",
-        destination: "https://skylab-backend.herokuapp.com/api/auth/info",
-      },
-      {
-        source: "/api/auth/sign-out",
-        destination: "https://skylab-backend.herokuapp.com/api/auth/sign-out",
-      },
-    ];
-  },
 };
 
 module.exports = nextConfig;
