+++ pages/sign-up.tsx
@@ -15,7 +15,6 @@ import { Cohort } from "@/types/cohorts";
 interface SignUpFormValuesType {
   name: string;
   email: string;
-  password: string;
   matricNo: string;
   nusnetId: string;
 }
@@ -31,7 +30,6 @@ const SignUp: NextPage = () => {
   const initialValues: SignUpFormValuesType = {
     name: "",
     email: "",
-    password: "",
     matricNo: "",
     nusnetId: "",
   };
@@ -40,7 +38,7 @@ const SignUp: NextPage = () => {
     values: SignUpFormValuesType,
     actions: FormikHelpers<SignUpFormValuesType>
   ) => {
-    const { name, email, password, matricNo, nusnetId } = values;
+    const { name, email, matricNo, nusnetId } = values;
     if (fetchCohortStatus === FETCH_STATUS.FETCHING) {
       alert("Still fetching latest cohort... Please try again");
       return;
@@ -49,7 +47,6 @@ const SignUp: NextPage = () => {
       await signUp({
         name,
         email,
-        password,
         matricNo,
         nusnetId,
         role: "students",
@@ -71,7 +68,6 @@ const SignUp: NextPage = () => {
         signUp({
           name: `Student ${i}`,
           email: `student${i}@gmail.com`,
-          password: "test1234",
           cohortYear,
           role: "students",
           matricNo: `A000000${i}X`,
@@ -88,7 +84,6 @@ const SignUp: NextPage = () => {
         signUp({
           name: `Adviser ${i}`,
           email: `adviser${i}@gmail.com`,
-          password: "test1234",
           cohortYear,
           role: "advisers",
         })
@@ -104,7 +99,6 @@ const SignUp: NextPage = () => {
         signUp({
           name: `Mentor ${i}`,
           email: `mentor${i}@gmail.com`,
-          password: "test1234",
           cohortYear,
           role: "mentors",
         })
@@ -147,12 +141,6 @@ const SignUp: NextPage = () => {
                       label="Email"
                       type="email"
                       name="email"
-                      formik={formik}
-                    />
-                    <TextInput
-                      label="Password"
-                      type="password"
-                      name="password"
                       formik={formik}
                     />
                     <TextInput
